Fix candidate passed as text, and optimize load
- When the candidate is passed as a text blob, split it into separate
lines, otherwise it'll attempt to load the entire blob as:

```
sudo config -s "config.alerts.migrated=on
config.auth.extendedsessionids on
config.auth.type Local
config.console.flow None
config.console.ppp.auth None

...
"
```

(notice that the ``=`` is placed only on the first line, therefore it
breaks the usage).

- As per
https://opengear.zendesk.com/hc/en-us/articles/216371123-Configuring-from-the-command-line,
under the _Hints & tips_, it is pointed out that

>  You can chain config commands, this can be significantly faster than
running separate commands as config.xml is only read/written once.

Not only that on the device itself it is faster to send only one
command, but also when executing hundreds of *config* commands over SSH
is obviously suboptimal (and also safer as if it fails, it fails
altogether rather than having loaded something already, which is more
in-line with the usual NAPALM methodology).